Use off_t, size_t, fseeko(), and ftello() for better 32-bit support.

+++ file.h
@@ -15,23 +15,23 @@ public:
     bool open(std::string filename);
     bool create(std::string filename);
 
-    void seek(int64_t p);
-    int64_t pos();
+    void seek(off_t p);
+    off_t pos();
     bool atEnd();
-    int64_t length() { return size; }
+    off_t length() { return size; }
 
     int readInt();
     int64_t readInt64();
-    void readChar(char *dest, int64_t n);
-    std::vector<unsigned char> read(int64_t n);
+    void readChar(char *dest, size_t n);
+    std::vector<unsigned char> read(size_t n);
 
     int writeInt(int n);
     int writeInt64(int64_t n);
-    int writeChar(char *source, int64_t n);
+    int writeChar(char *source, size_t n);
     int write(std::vector<unsigned char> &v);
 
 protected:
-    int64_t size;
+    off_t size;
     FILE *file;
 };
